Provides Extensive Comments On The Codebase With Programming-By-Contract (PBC) And Test-Driven Development (TDD) Techniques. We Utilize PyTest For Testing. The Tests Are Located In The Tests Directory. The Tests Are Run By Typing "pytest" In The Terminal With Specified Modes To Select Varying Means Of Testing LIke Unit, Integration, And Functional Testing.
Also Implemented A More Lavish Front-End With PyQT5. The Front-End Is Located In The src/front-end Directory. The Front-End
Is Run By Typing "python client_window.py" In The Terminal. The Front-End Is Used To Interact With The User And Display
The Results Of The Analysis As Well As Allow The User To Change The Settings Of The Analysis.

A New Gap Detection Algorithm Was Implemented. The Algorithm Is Located In The src/gap_detection Directory. The Algorithm
Is Run By Typing "python gap_detection_demo.py" In The Terminal. The Algorithm Is Used To Detect Gaps In The Canopy Of
The Forest. The Algorithm Is Used To Detect Gaps In The Canopy Of The Forest By Utilizing A Depth Estimation Model To
Estimate The Depth Of The Canopy. The Depth Estimation Model Is Trained On A Large Dataset Of Images Of The Canopy.